+++ src/commands/hardis/project/clean/minimizeprofiles.ts
@@ -91,6 +91,10 @@ You can override this list by defining a property minimizeProfilesNodesToRemove 
     }
 
     // Summary
+    if (counter > 0) {
+      uxLog(this, c.yellow("Please make sure the attributes removed from Profiles are defined on Permission Sets"));
+      globalThis.displayProfilesWarning = true;
+    }
     const msg = `Cleaned ${c.green(c.bold(counter))} profiles from attributes existing on Permission Sets`;
     uxLog(this, c.cyan(msg));
     // Return an object to be displayed with --json