+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+## [4.18.2] 2023-11-29
+
+- **hardis:work:save** enhancements
+  - Display more output during cleaning jobs
+  - Keep **userPermissions** in Profiles when they are defined to `false`
+
 ## [4.18.1] 2023-11-29
 
 - Improve backup notifications display